requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.9
Django==5.1.4
django-allauth==65.3.1
django-cleanup==9.0.0
django-htmx==1.21.0
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
identify==2.6.4
nodeenv==1.9.1
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
virtualenv==20.28.0
